Allow parameters to be pulled from YAML files.
Sometimes you might be making a series of modules all with similar
parameters.
It's usually a lot easier to copy and edit a file globally than to try
and do these edits on the command line.
You might also have a wrapper script that always sources a specific file
on your filesystem.